dialect namespace
